fix dtype in random shuffle

+++ roughpy_jax/tests/test_ft_mul_adj.py
@@ -1,4 +1,3 @@
-
 import pytest
 
 import jax
@@ -9,9 +8,7 @@ from numpy.testing import assert_array_almost_equal, assert_array_equal
 import roughpy_jax as rpj
 
 
-
 def test_adjoint_left_ft_mul_identity():
-
     basis = rpj.TensorBasis(2, 2)
 
     A_data = jnp.array([1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], dtype=jnp.float32)
@@ -20,7 +17,8 @@ def test_adjoint_left_ft_mul_identity():
 
     rng = jax.random.key(12345)
     S = rpj.DenseShuffleTensor(
-        jax.random.uniform(rng, minval=-1.0, maxval=1.0, shape=(basis.size(),)),
+        jax.random.uniform(rng, minval=-1.0, maxval=1.0, dtype=jnp.float32,
+                           shape=(basis.size(),)),
         basis
     )
 
@@ -38,12 +36,14 @@ def test_adjoint_left_ft_mul_letter():
 
     rng = jax.random.key(12345)
     S = rpj.DenseShuffleTensor(
-        jax.random.uniform(rng, minval=-1.0, maxval=1.0, shape=(basis.size(),)),
+        jax.random.uniform(rng, minval=-1.0, maxval=1.0, dtype=jnp.float32,
+                           shape=(basis.size(),)),
         basis
     )
 
     R = rpj.ft_adjoint_left_mul(A, S)
 
-    expected_data = jnp.array([S.data[1], S.data[3], S.data[4], 0.0, 0.0, 0.0, 0.0])
+    expected_data = jnp.array(
+        [S.data[1], S.data[3], S.data[4], 0.0, 0.0, 0.0, 0.0])
 
     assert_array_equal(R.data, expected_data)
